Move alliance_formation into a new examples/experimental module
The alliance formation example is built on `mesa.experimental.meta_agents`,
which carries no semver guarantees, so it doesn't belong alongside the
advanced examples. Add a third `mesa/examples/experimental/` group and move
the example there in full.

Also fixes the package's `__init__ .py` (leading space), which meant the
example directory had no real `__init__.py`, and adds the example to
examples/README.md under a new Experimental section.
