+++ mesa/examples/experimental/alliance_formation/agents.py
